ResolveError

+++ src/data.rs
@@ -1,6 +1,6 @@
 use cgmath::{Vector3, Vector2};
 
-use crate::{Pref, CHUNK_SIZE};
+use crate::{CHUNK_SIZE};
 
 pub struct WorldData {
     pub basic_data: Vec<VoxelChunk>,
@@ -33,7 +33,7 @@ impl WorldData {
     pub fn collect() -> WorldData {
         // Remove Later
         let mut basic_voxel_input: Vec<u32> = vec![];
-        for index in 0 .. CHUNK_SIZE { basic_voxel_input.push(index); }
+        for index in 0 .. CHUNK_SIZE { basic_voxel_input.push(index as u32); }
 
         let voxel_chunk = VoxelChunk { voxel_data: WorldData::vec_to_array(basic_voxel_input) };
         WorldData { basic_data: vec![voxel_chunk] }
